requirements.txt
Flask
gunicorn
celery
redis
requests
google-generativeai
python-dotenv
gevent